requirements.txt
ebyte-lora-e22-rpi
RPi.GPIO
pyserial
python-dotenv
paho.mqtt